+++ app/src/main/java/com/dan/school/SplashScreenActivity.kt
@@ -5,9 +5,6 @@ import android.content.Intent
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import com.dan.school.setup.SetupActivity
-import com.google.firebase.auth.FirebaseAuth
-import com.google.firebase.auth.ktx.auth
-import com.google.firebase.ktx.Firebase
 
 class SplashScreenActivity : AppCompatActivity() {
