--- app/src/main/java/com/dan/school/interfaces/BackupItemClickListener.kt
@@ -1,8 +1,0 @@
-package com.dan.school.interfaces
-
-import com.google.firebase.storage.StorageReference
-
-/** Invoked when a RecyclerView item is clicked */
-interface BackupItemClickListener {
-    fun backupItemClicked(storageReference: StorageReference)
-}
