+++ app/src/main/java/com/dan/school/settings/ProfileFragment.kt
@@ -2,7 +2,6 @@ package com.dan.school.settings
 
 import android.animation.LayoutTransition
 import android.content.Context
-import android.content.Intent
 import android.content.SharedPreferences
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -12,24 +11,13 @@ import android.view.inputmethod.InputMethodManager
 import android.widget.Toast
 import androidx.activity.OnBackPressedCallback
 import androidx.core.content.edit
-import androidx.core.view.isGone
 import androidx.core.view.isVisible
 import androidx.fragment.app.Fragment
 import com.dan.school.ProgressBarDialog
 import com.dan.school.R
 import com.dan.school.School
-import com.dan.school.Utils
-import com.dan.school.authentication.AuthenticationActivity
 import com.dan.school.databinding.FragmentProfileBinding
 import com.google.android.material.dialog.MaterialAlertDialogBuilder
-import com.google.firebase.auth.FirebaseAuth
-import com.google.firebase.auth.ktx.auth
-import com.google.firebase.database.FirebaseDatabase
-import com.google.firebase.database.ServerValue
-import com.google.firebase.database.ktx.database
-import com.google.firebase.ktx.Firebase
-import java.lang.Exception
-import kotlin.math.ceil
 
 class ProfileFragment : Fragment() {
 
@@ -37,13 +25,10 @@ class ProfileFragment : Fragment() {
 
     private val binding get() = _binding!!
 
-    private var isSendingVerificationEmail = false
     private var isEditMode = false
     private lateinit var sharedPref: SharedPreferences
 
     private lateinit var inputMethodManager: InputMethodManager
-    private lateinit var auth: FirebaseAuth
-    private lateinit var database: FirebaseDatabase
 
     private lateinit var progressBarDialog: ProgressBarDialog
 
@@ -74,9 +59,6 @@ class ProfileFragment : Fragment() {
                 }
             }
         requireActivity().onBackPressedDispatcher.addCallback(this, callback)
-
-        auth = Firebase.auth
-        database = Firebase.database
     }
 
     private fun showSaveChangesDialog() {
@@ -106,12 +88,6 @@ class ProfileFragment : Fragment() {
         return binding.root
     }
 
-    override fun onResume() {
-        super.onResume()
-        binding.swipeRefreshLayout.isRefreshing = true
-        update()
-    }
-
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
@@ -119,10 +95,6 @@ class ProfileFragment : Fragment() {
 
         binding.textViewNicknameDisplay.text = sharedPref.getString(School.NICKNAME, "")
         binding.textViewFullNameDisplay.text = sharedPref.getString(School.FULL_NAME, "")
-
-        binding.buttonSendVerificationEmail.setOnClickListener {
-            buttonSendVerificationEmailClicked()
-        }
 
         binding.buttonEdit.setOnClickListener {
             if (isEditMode) {
@@ -137,220 +109,11 @@ class ProfileFragment : Fragment() {
         binding.buttonBack.setOnClickListener {
             requireActivity().onBackPressed()
         }
-
-        binding.buttonResetPassword.setOnClickListener {
-            val time = Utils.getDifferenceInSeconds(
-                sharedPref.getLong(
-                    School.PASSWORD_RESET_EMAIL_TIME_LAST_SENT,
-                    0
-                )
-            )
-            if (time >= 30) {
-                showResetPasswordDialog()
-            } else {
-                val timeToWait = ceil(30 - time).toInt()
-                Toast.makeText(
-                    requireContext(),
-                    "${getString(R.string.please_try_again_in)} $timeToWait ${
-                        if (timeToWait == 1) getString(
-                            R.string.second
-                        ) else getString(R.string.seconds)
-                    }.",
-                    Toast.LENGTH_SHORT
-                ).show()
-            }
-        }
-
-        binding.buttonSignIn.setOnClickListener {
-            val intent = Intent(requireContext(), AuthenticationActivity::class.java)
-            intent.putExtra(School.SHOW_BUTTON_SIGN_IN_LATER, false)
-            startActivity(intent)
-        }
-
-        binding.swipeRefreshLayout.setOnRefreshListener {
-            update()
-        }
-    }
-
-    private fun showResetPasswordDialog() {
-        MaterialAlertDialogBuilder(requireContext()).setMessage("${getString(R.string.send_password_reset_email_to)} ${auth.currentUser?.email}.")
-            .setTitle("${getString(R.string.reset_password)}?")
-            .setPositiveButton(
-                getString(R.string.yes)
-            ) { _, _ ->
-                showProgressBar()
-                auth.currentUser?.email?.let { email ->
-                    sendPasswordResetEmail(email)
-                }
-            }
-            .setNegativeButton(
-                getString(R.string.no)
-            ) { _, _ -> }
-            .create()
-            .show()
-    }
-
-    private fun sendPasswordResetEmail(email: String) {
-        auth.sendPasswordResetEmail(email).addOnCompleteListener {
-            if (it.isSuccessful) {
-                try {
-                    Toast.makeText(
-                        requireContext(),
-                        getString(R.string.password_reset_email_sent),
-                        Toast.LENGTH_SHORT
-                    ).show()
-                } catch (e: Exception) {
-                }
-                savePasswordResetEmailTimeLastSent()
-            } else {
-                try {
-                    Toast.makeText(
-                        requireContext(),
-                        getString(R.string.an_error_occurred),
-                        Toast.LENGTH_SHORT
-                    ).show()
-                } catch (e: Exception) {
-                }
-            }
-            hideProgressBar()
-        }
-    }
-
-    private fun savePasswordResetEmailTimeLastSent() {
-        with(sharedPref.edit()) {
-            putLong(
-                School.PASSWORD_RESET_EMAIL_TIME_LAST_SENT,
-                System.currentTimeMillis()
-            )
-            apply()
-        }
     }
 
     override fun onDestroyView() {
         super.onDestroyView()
         _binding = null
-    }
-
-    private fun showProgressBar() {
-        progressBarDialog.show()
-    }
-
-    private fun hideProgressBar() {
-        progressBarDialog.dismiss()
-    }
-
-    /**
-     * Updates the visibility of [FragmentProfileBinding.groupEmail],
-     * [FragmentProfileBinding.buttonResetPassword],
-     * [FragmentProfileBinding.cardViewVerifyEmail] and
-     * [FragmentProfileBinding.buttonSignIn] depending if a
-     * user is signed-in and if the email is verified.
-     * Sets the text of [FragmentProfileBinding.textViewEmailDisplay]
-     * to the current user's email.
-     */
-    private fun update() {
-        val user = auth.currentUser
-        if (user != null) {
-
-            binding.buttonSignIn.visibility = View.GONE
-
-            binding.groupEmail.visibility = View.VISIBLE
-            binding.buttonResetPassword.visibility = View.VISIBLE
-
-            binding.textViewEmailDisplay.text = user.email
-            binding.cardViewVerifyEmail.isGone = user.isEmailVerified
-
-            if (!user.isEmailVerified) {
-                user.reload().addOnCompleteListener {
-                    if (_binding == null) {
-                        return@addOnCompleteListener
-                    }
-                    if (it.isSuccessful) {
-                        binding.textViewEmailDisplay.text = user.email
-                        binding.cardViewVerifyEmail.isGone = user.isEmailVerified
-                    } else {
-                        try {
-                            Toast.makeText(
-                                requireContext(),
-                                getString(R.string.failed_to_update),
-                                Toast.LENGTH_SHORT
-                            ).show()
-                        } catch (e: Exception) {
-                        }
-                    }
-                    binding.swipeRefreshLayout.isRefreshing = false
-                }
-            } else {
-                binding.swipeRefreshLayout.isRefreshing = false
-            }
-        } else {
-            binding.groupEmail.visibility = View.GONE
-            binding.buttonResetPassword.visibility = View.GONE
-            binding.buttonSignIn.visibility = View.VISIBLE
-            binding.cardViewVerifyEmail.isGone = true
-            binding.swipeRefreshLayout.isRefreshing = false
-        }
-    }
-
-    /**
-     * Sends email verification on currently signed-in user if
-     * the last time a verification email was sent is 30 seconds
-     * before [System.currentTimeMillis].
-     */
-    private fun buttonSendVerificationEmailClicked() {
-        val time = Utils.getDifferenceInSeconds(
-            sharedPref.getLong(
-                School.VERIFICATION_EMAIL_TIME_LAST_SENT,
-                0
-            )
-        )
-        if (time >= 30) {
-            if (!isSendingVerificationEmail) {
-                isSendingVerificationEmail = true
-                binding.cardViewVerifyEmail.visibility = View.VISIBLE
-                sendEmailVerification()
-            }
-        } else {
-            val timeToWait = ceil(30 - time).toInt()
-            Toast.makeText(
-                requireContext(),
-                "${getString(R.string.please_try_again_in)} $timeToWait ${
-                    if (timeToWait == 1) getString(
-                        R.string.second
-                    ) else getString(R.string.seconds)
-                }.",
-                Toast.LENGTH_SHORT
-            ).show()
-        }
-    }
-
-    private fun sendEmailVerification() {
-        auth.currentUser?.sendEmailVerification()?.addOnCompleteListener {
-            if (it.isSuccessful) {
-                with(sharedPref.edit()) {
-                    putLong(
-                        School.VERIFICATION_EMAIL_TIME_LAST_SENT,
-                        System.currentTimeMillis()
-                    )
-                    apply()
-                }
-                Toast.makeText(
-                    requireContext(),
-                    getString(R.string.verification_email_sent),
-                    Toast.LENGTH_SHORT
-                ).show()
-            } else {
-                try {
-                    Toast.makeText(
-                        requireContext(),
-                        getString(R.string.failed_to_send_verification_email),
-                        Toast.LENGTH_SHORT
-                    ).show()
-                } catch (e: Exception) {
-                }
-            }
-            isSendingVerificationEmail = false
-        }
     }
 
     private fun setEditMode(editMode: Boolean): Boolean {
@@ -394,19 +157,7 @@ class ProfileFragment : Fragment() {
         sharedPref.edit {
             putString(School.NICKNAME, nickname)
             putString(School.FULL_NAME, fullName)
-            sharedPref.edit().putBoolean(School.DATABASE_PROFILE_UPDATED, false).apply()
             commit()
-        }
-        auth.currentUser?.let {
-            val map = mapOf(
-                School.NICKNAME to nickname,
-                School.FULL_NAME to fullName,
-                School.PROFILE_LAST_UPDATE_TIME to ServerValue.TIMESTAMP
-            )
-            database.reference.child(School.USERS).child(it.uid).updateChildren(map)
-                .addOnSuccessListener {
-                    sharedPref.edit().putBoolean(School.DATABASE_PROFILE_UPDATED, true)
-                }
         }
     }
