--- app/src/main/res/layout/fragment_authentication.xml
@@ -1,73 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:gravity="center"
-    android:orientation="vertical"
-    tools:context=".authentication.AuthenticationFragment">
-
-    <TextView
-        android:id="@+id/textViewSignIn"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="36dp"
-        android:layout_marginEnd="36dp"
-        android:fontFamily="@font/cabin_bold"
-        android:text="@string/sign_in"
-        android:textColor="@color/colorPrimary"
-        android:textSize="48sp" />
-
-    <com.google.android.material.button.MaterialButton
-        android:id="@+id/buttonSignInWithEmail"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="36dp"
-        android:layout_marginTop="16dp"
-        android:layout_marginEnd="36dp"
-        android:paddingLeft="0dp"
-        android:paddingTop="0dp"
-        android:paddingBottom="0dp"
-        android:text="@string/sign_in_with_email"
-        android:textColor="@color/signInButtonTextColor"
-        app:backgroundTint="@color/signInWithEmailBackgroundColor"
-        app:icon="@drawable/ic_sign_in_with_email"
-        app:iconPadding="12dp"
-        app:iconSize="40dp"
-        app:iconTint="?android:textColorPrimary"
-        tools:ignore="RtlHardcoded,RtlSymmetry" />
-
-    <com.google.android.material.button.MaterialButton
-        android:id="@+id/buttonSignInWithGoogle"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="36dp"
-        android:layout_marginEnd="36dp"
-        android:paddingLeft="0dp"
-        android:paddingTop="0dp"
-        android:paddingBottom="0dp"
-        android:text="@string/sign_in_with_google"
-        android:textColor="@color/signInButtonTextColor"
-        app:backgroundTint="@color/signInWithGoogleBackgroundColor"
-        app:icon="@drawable/ic_google"
-        app:iconPadding="12dp"
-        app:iconSize="40dp"
-        app:iconTint="@null"
-        tools:ignore="RtlHardcoded,RtlSymmetry" />
-
-    <com.google.android.material.button.MaterialButton
-        android:id="@+id/buttonSignInLater"
-        style="@style/Widget.MaterialComponents.Button.TextButton"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="Sign in later" />
-
-    <com.google.android.material.button.MaterialButton
-        android:id="@+id/buttonSignUp"
-        style="@style/Widget.MaterialComponents.Button.TextButton"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="@string/no_account_yet_sign_up" />
-
-</LinearLayout>
